requirements.txt
flask
pyyaml
nltk
numpy
SpeechRecognition
metaphone
nltk.download('wordnet')